store/tikv/mocktikv: remove the old code
