+++ store/tikv/mock-tikv/topn.go
@@ -138,30 +138,3 @@ func (t *topnHeap) tryToAddRow(row *sortRow) bool {
 	}
 	return success
 }
-
-// evalTopN evaluates the top n elements from the data. The input receives a record including its handle and data.
-// And this function will check if this record can replace one of the old records.
-func (h *rpcHandler) evalTopN(ctx *selectContext, handle int64, values map[int64][]byte, columns []*tipb.ColumnInfo) error {
-	err := setColumnValueToEval(ctx.eval, handle, values, ctx.topnColumns)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	newRow := &sortRow{
-		meta: tipb.RowMeta{Handle: handle},
-	}
-	for _, item := range ctx.topnHeap.orderByItems {
-		result, err := ctx.eval.Eval(item.Expr)
-		if err != nil {
-			return errors.Trace(err)
-		}
-		newRow.key = append(newRow.key, result)
-	}
-	if ctx.topnHeap.tryToAddRow(newRow) {
-		for _, col := range columns {
-			val := values[col.GetColumnId()]
-			newRow.data = append(newRow.data, val...)
-			newRow.meta.Length += int64(len(val))
-		}
-	}
-	return errors.Trace(ctx.topnHeap.err)
-}
